retrieve the images stored from the database...

+++ docker-compose/mysql/initialization_scripts/init.sql
@@ -2,7 +2,7 @@
 CREATE TABLE images(
    transaction_id INT NOT NULL AUTO_INCREMENT,
    image_id VARCHAR(100) NOT NULL,
-   scale INT NOT NULL,
+   scale DECIMAL NOT NULL,
    image_url VARCHAR(256) NOT NULL,
    PRIMARY KEY ( transaction_id )
 );
